// types/TmdbMovie.ts
type TmdbMovie = {
  poster_path?: string | null
  adult?: boolean
  overview?: string
  release_date?: string
  genre_ids?: [number]
  id: number // NOTE: Assuming that TMDB is giving an ID
  original_title?: string
  original_language?: string
  title?: string
  backdrop_path?: string | null
  popularity?: number
  vote_count?: number
  video?: boolean
  vote_average?: number
}

export default TmdbMovie